size_t is is dependent on the underlying system (32-bit, 64 bit) and thus may not be used

+++ streamconnector/workflowserializerv1.cpp
@@ -358,7 +358,7 @@ void WorkflowSerializerV1::loadNode(SPWorkFlowNode& node,Workflow *workflow, con
         bool quotes;
         _stream >> quotes;
         wp.needsQuotes(quotes);
-        size_t sz;
+        qint32 sz;
         _stream >> sz;
         std::vector<XY> line;
         for(int p=0; p < sz; ++p){
